autodevcrew1/requirements.txt
# Core Dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
autogen>=0.2.0
streamlit>=1.28.0
pytest>=7.4.0
openai>=1.0.0
pydantic>=2.0.0

# Privacy & Security
cryptography>=41.0.0

# Hardware Optimization
# torch>=2.0.0
psutil>=5.9.0
# llama-cpp-python>=0.2.0

# Cloud Deployment
huggingface-hub>=0.19.0
pyngrok>=7.0.0

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0

# Configuration & Utilities
PyYAML>=6.0.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# GitHub Integration
PyGithub>=2.1.1

# Monitoring & Logging
prometheus-client>=0.19.0

# Additional utilities
aiohttp>=3.9.0
python-multipart>=0.0.6
